// src/app/models/paymentVoucher.ts
export class PaymentVoucher {
    id?: number; // Optional, as the ID may be generated by the backend
    voucherNo: string;
    restaurantName: string;
    paymentDate: Date;
    invoiceFor: string;
    lunchDate: Date;
    date: Date;
    pax: number;
    price: number;
    amount: number; // pax * price
    remark: string;
    cashier: string;
    receivedBy: string;
    approvedBy: string;
  
    constructor(
      id: number = 0,
      voucherNo: string = '',
      restaurantName: string = '',
      paymentDate: Date = new Date(),
      invoiceFor: string = '',
      lunchDate: Date = new Date(),
      date: Date = new Date(),
      pax: number = 0,
      price: number = 0,
      amount: number = 0,
      remark: string = '',
      cashier: string = '',
      receivedBy: string = '',
      approvedBy: string = ''
    ) {
      this.id = id;
      this.voucherNo = voucherNo;
      this.restaurantName = restaurantName;
      this.paymentDate = paymentDate;
      this.invoiceFor = invoiceFor;
      this.lunchDate = lunchDate;
      this.date = date;
      this.pax = pax;
      this.price = price;
      this.amount = amount;
      this.remark = remark;
      this.cashier = cashier;
      this.receivedBy = receivedBy;
      this.approvedBy = approvedBy;
    }
  }
  